error de subir archivo actualizacoin arreglado

+++ src/controllers/autosController.js
@@ -59,21 +59,28 @@ const controlador =
         let imagenABorrar;
 		for (let s of productos){
 			if(idProducto==s.id){   
-                imagenABorrar= s.imagen;
+                
 				s.modelo= req.body.modelo,
 				s.categoria= req.body.category,
                 s.ano= req.body.ano,
 				s.precio= req.body.precio,
                 s.fechaDispDesde= req.body.fechaDispDesde,
-                s.fechaDispHasta = req.body.fechaDispHasta, 
-                s.imagen= req.file.filename
-               break;
+                s.fechaDispHasta = req.body.fechaDispHasta 
+                if(req.file != undefined){
+                    imagenABorrar= s.imagen;
+                    s.imagen=req.file.filename
+                    fs.unlinkSync(path.join(__dirname, '../../public/img/', imagenABorrar));
+                } else {
+                    s.imagen =s.imagen
+                    
+                }
+                break;
             }}    
+
         fs.writeFileSync(productosFilePath, JSON.stringify(productos, null, " ")),
         console.log(imagenABorrar)
+        
                    
-        fs.unlinkSync(path.join(__dirname, '../../public/img/', imagenABorrar));
-
         res.redirect('/');  
         
     },
